Working example for explainer
The logging in the explaioner example previously ran immediately instead
of only on success, as it wasn't placed in a callback, i.e. the
'() => {' and '}' was missing.

Also, the example previously caused an exception if the user declined
to proceed with the share.

The explainer now uses the working example from the spec.