Optional message to display when sharing
For sites that wish to display a message when the
share dialog is presented, we add a new ShareOptions
dictionary.

This allows for additional options that may be needed in future,
without long method parameter lists.

Note that support for the 'message' option (and likely future
options) is not feature detectable.