Check transient activation after sharePromise
WPT web-share/share-sharePromise-internal-slot.https.html
has multiple navigator.share() calls within a single
onclick handler. This is a good way to test the sharePromise
internal slot, but it was not previously permitted by the spec,
as the first call would consume activation and the later calls
would fail with NotAllowedError instead of InvalidStateError.